pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "byte-brewery"
version = "1.0.0"
description = "Scripts crafted like fine, magical ales for AI development"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.75.0",
    "google-genai>=1.56.0",
    "openai>=2.14.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-watch>=4.2.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
# Main Pipeline
aug = "augment_ai.aug_pipeline:main"

# Utilities (Now properly namespaced)
aug-recap = "augment_ai.recap:main"
dce = "augment_ai.dce:main"
pext = "augment_ai.pext:main"
aireview = "aireview.main:main"

[tool.setuptools.packages.find]
where = ["src"]
